refactor(app): extract Editor constructors and init (phase 6e)
Move Editor::new, Editor::with_working_dir, Editor::for_test,
Editor::with_options (the mega-constructor that initializes every field
on Editor, ~600 lines), and start_background_grammar_build out of
mod.rs into a new app/editor_init.rs sibling. ~770 lines.

The constructor cluster touches every type Editor uses, so editor_init.rs
deliberately uses `use super::*` to inherit mod.rs's existing import set
rather than repeat them. start_background_grammar_build becomes
pub(super) so the deferred-grammar-build path in mod.rs can call it.

mod.rs goes from 4882 -> 4126 lines (-756). 2026 tests still pass.
